+++ packages/keycloak-user-management/src/components/UserGroupDetailView/index.tsx
@@ -1,25 +1,24 @@
 import React from 'react';
-import { Col, Space, Typography } from 'antd';
+import { Col, Space, Typography, Spin } from 'antd';
 import { CloseOutlined } from '@ant-design/icons';
 import { useHistory } from 'react-router';
 import { Resource404 } from '@opensrp/react-utils';
+import { sendErrorNotification } from '@opensrp/notifications';
 import { Button } from 'antd';
 import { URL_USER_EDIT, URL_USER_GROUPS } from '../../constants';
 import { UserGroupMembers } from '../UserGroupsList';
 import { KeycloakUserGroup } from 'keycloak-user-management/src/ducks/userGroups';
 import { Link } from 'react-router-dom';
+import { loadGroupDetails, loadGroupMembers } from '../UserGroupsList/utils';
 const { Text } = Typography;
 
 /** typings for the view details component */
 export interface ViewDetailsProps {
-  singleUserGroupDetails: KeycloakUserGroup | null;
   groupId: string;
-  userGroupMembers: UserGroupMembers[] | null;
+  keycloakBaseURL: string;
 }
 export const defaultProps = {
   groupId: '',
-  singleUserGroupDetails: null,
-  userGroupMembers: null,
 };
 /** component that renders the details view to the right side
  * of list view
@@ -27,11 +26,28 @@ export const defaultProps = {
  * @param props - detail view component props
  */
 const ViewDetails = (props: ViewDetailsProps) => {
-  const { groupId, userGroupMembers, singleUserGroupDetails } = props;
+  const { groupId, keycloakBaseURL } = props;
   const history = useHistory();
+  const [loading, setLoading] = React.useState<boolean>(true);
+  const [userGroupMembers, setUserGroupMembers] = React.useState<UserGroupMembers[] | null>(null);
+  const [singleUserGroup, setSingleUserGroup] = React.useState<KeycloakUserGroup | null>(null);
+
+  React.useEffect(() => {
+    if (groupId) {
+      setLoading(true);
+      const membersPromise = loadGroupMembers(groupId, keycloakBaseURL, setUserGroupMembers);
+      const userGroupPromise = loadGroupDetails(groupId, keycloakBaseURL, setSingleUserGroup);
+      Promise.all([membersPromise, userGroupPromise])
+        .catch((e) => sendErrorNotification(`${e}`))
+        .finally(() => setLoading(false));
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [groupId]);
+
   if (groupId === '') {
     return null;
   }
+
   return (
     <Col className="view-details-content">
       <div className="flex-right">
@@ -41,7 +57,10 @@ const ViewDetails = (props: ViewDetailsProps) => {
           onClick={() => history.push(URL_USER_GROUPS)}
         />
       </div>
-      {groupId && (!singleUserGroupDetails || !userGroupMembers) ? (
+      {loading ? (
+        <Spin size="large" className="custom-ant-spin" />
+      ) : // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+      !loading && groupId && (!singleUserGroup || !userGroupMembers) ? (
         <Resource404 />
       ) : (
         <Space direction="vertical">
@@ -49,12 +68,12 @@ const ViewDetails = (props: ViewDetailsProps) => {
             Name
           </Text>
           <Text type="secondary" className="display-block">
-            {singleUserGroupDetails?.name}
+            {singleUserGroup?.name}
           </Text>
           <Text strong={true} className="display-block">
             Roles
           </Text>
-          {singleUserGroupDetails?.realmRoles?.map((role: string) => (
+          {singleUserGroup?.realmRoles?.map((role: string) => (
             <Text key={role} type="secondary" className="display-block">
               {role}
             </Text>
